requirements.txt

streamlit
pillow
    